requirements.txt
# PyTorch with CUDA 11.8 support (GTX 1650 compatible)
--index-url https://download.pytorch.org/whl/cu118
torch==2.1.0+cu118
torchvision==0.16.0+cu118

# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Computer Vision and Data Processing
opencv-python-headless==4.8.1.78
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# Visualization and Utilities
matplotlib==3.8.2
tqdm==4.66.1
pillow==10.1.0
